+++ src/components/IFrame.js
@@ -16,18 +16,21 @@ type Props = {
   onChargeFailure: (MessageData) => void,
   onPaymentDetected: (MessageData) => void,
   onError: (MessageData) => void,
-  onModalClose: () => void
+  onModalClose: () => void,
+  disableCaching: boolean
 };
 
 type State = {
-  loading: boolean
+  loading: boolean,
+  src: null | string
 };
 
 type SrcParams = {
   origin: string,
   version: string,
   buttonId: string,
-  custom?: string
+  custom?: string,
+  cacheDisabled: boolean
 }
 
 export default class IFrame extends React.Component<Props, State> {
@@ -40,24 +43,26 @@ export default class IFrame extends React.Component<Props, State> {
 
     this.origin = 'https://commerce.coinbase.com';
     this.uuid = generateUUID();
-    this.hostName = `${window.location.protocol}//${window.location.hostname}${window.location.port ? ':' + window.location.port : ''}`;
 
     this.state = {
-      loading: true
+      loading: true,
+      src: null
     }
   }
 
   componentDidMount(){
     // Add event listeners for the iframe
     window.addEventListener('message', this.handleMessage);
+    const hostName = `${window.location.protocol}//${window.location.hostname}${window.location.port ? ':' + window.location.port : ''}`;
+    this.setState({ src: this.buildSrc(hostName) });
   }
 
   componentWillUnmount() {
     window.removeEventListener('message', this.handleMessage);
   }
 
-  buildSrc = (): string => {
-    const {checkoutId, chargeId, customMetadata} = this.props;
+  buildSrc = (hostName: string): string => {
+    const {checkoutId, chargeId, customMetadata, disableCaching} = this.props;
 
     function encodeURIParams(params) {
       let encoded = [];
@@ -83,9 +88,10 @@ export default class IFrame extends React.Component<Props, State> {
     }
 
     const params: SrcParams = {
-      origin: this.hostName,
+      origin: hostName,
       version: VERSION,
       buttonId: this.uuid,
+      cacheDisabled: disableCaching
     };
 
     let custom = '';
@@ -151,20 +157,22 @@ export default class IFrame extends React.Component<Props, State> {
   };
 
   render() {
-    const src = this.buildSrc();
+    const { loading, src } = this.state;
+
     return (
       <div className="coinbase-commerce-iframe-container">
-        {this.state.loading ? (
+        {loading || src === null && (
           <div className="commerce-loading-spinner"/>
-        ) : (null)}
-        <iframe
-          onLoad={this.handleIFrameLoaded}
-          className="coinbase-commerce-iframe"
-          src={src}
-          allowtransparency={'yes'}
-          scrolling={'no'}
-          frameBorder="no"
-        />
+        )}
+        {src !== null && (
+          <iframe
+            onLoad={this.handleIFrameLoaded}
+            className="coinbase-commerce-iframe"
+            src={src}
+            scrolling="no"
+            frameBorder="no"
+          />
+        )}
       </div>
     )
   }