+++ src/index.js
@@ -14,8 +14,11 @@ type Props = {
   onChargeSuccess?: (MessageData) => void,
   onChargeFailure?: (MessageData) => void,
   onPaymentDetected?: (MessageData) => void,
-  onModalClosed?: () => void
+  onModalClosed?: () => void,
+  disableCaching: true,
+  wrapperStyle?: { [string]: number | string }
 };
+
 type State = {
   showModal: boolean
 };
@@ -29,7 +32,9 @@ const getButtonProps = (props) => {
     'customMetadata',
     'onPaymentDetected',
     'checkoutId',
-    'chargeId'
+    'chargeId',
+    'disableCaching',
+    'wrapperStyle'
   ];
   ignoredProps.forEach(p => delete buttonProps[p]);
   return buttonProps;
@@ -66,12 +71,14 @@ class CoinbaseCommerceButton extends React.Component<Props, State>{
 
   render(){
     const {showModal} = this.state;
-    const {onLoad, onChargeSuccess, onChargeFailure, checkoutId, chargeId, customMetadata, onPaymentDetected} = this.props;
-    const iFrameProps = {onLoad, onChargeSuccess, onChargeFailure, checkoutId, chargeId, onPaymentDetected};
+    const {onLoad, onChargeSuccess, onChargeFailure, checkoutId, chargeId, customMetadata, onPaymentDetected, disableCaching, wrapperStyle} = this.props;
+    const iFrameProps = {onLoad, onChargeSuccess, onChargeFailure, checkoutId, chargeId, onPaymentDetected, disableCaching};
     const buttonProps = getButtonProps(this.props);
     return (
-      <div>
-        <Button {...buttonProps} onClick={this.handleButtonClick}/>
+      <div style={wrapperStyle}>
+        <a href="https://commerce.coinbase.com" rel="external" title="Pay with Bitcoin, Bitcoin Cash, Litecoin, or Ethereum" onClick={e => e.preventDefault()}>
+          <Button {...buttonProps} onClick={this.handleButtonClick}/>
+        </a>
         {showModal && (
           <IFrame {...iFrameProps} onModalClose={this.handleModalClose} onError={this.handleError} customMetadata={customMetadata}/>
         )}