+++ CHANGELOG.md
@@ -1,8 +1,21 @@
 # CHANGELOG
 All notable changes to this project will be documented in this file.
 
-## 1.1.0 - 2018-09-17
+## 1.3.1 - 2018-10-22
 ### Changed
+- Removed obsolete `allowTransparency` prop from iframe (fixes [#15](https://github.com/coinbase/react-coinbase-commerce/issues/15))
+
+## 1.3.0 - 2018-10-22
+### Added
+- Caching the state of a user's transaction can now be disabled with the `disableCaching` prop.
+- The component is now wrapped in an anchor tag
+
+## 1.2.0 - 2018-10-01
+### Added
+- You can now register for `onPaymentDetected` events.
+
+## 1.1.0 - 2018-09-17
+### Added
 - Allow custom metadata to be included with charges
 
 ## 1.0.2 - 2018-09-13